test: move common test utils to top-level internal/
These helper functions are well-suited
for testing extensions, which requires
mocking schema.Cell and schema.Notebook.
