+++ render/html/html_test.go
@@ -9,9 +9,9 @@ import (
 	"github.com/google/go-cmp/cmp"
 	"github.com/stretchr/testify/require"
 
+	"github.com/bevzzz/nb/internal/test"
 	"github.com/bevzzz/nb/render"
 	"github.com/bevzzz/nb/render/html"
-	"github.com/bevzzz/nb/render/internal/test"
 	"github.com/bevzzz/nb/schema"
 )
