+++ render/html/wrapper_test.go
@@ -11,8 +11,8 @@ import (
 
 	"github.com/stretchr/testify/require"
 
+	"github.com/bevzzz/nb/internal/test"
 	"github.com/bevzzz/nb/render/html"
-	"github.com/bevzzz/nb/render/internal/test"
 	"github.com/bevzzz/nb/schema"
 	"github.com/bevzzz/nb/schema/common"
 )
